added valid POST req | POST ui and auth remaining
