+++ app/api/lab/route.ts
@@ -1,4 +1,6 @@
 import { db } from '@/lib/prisma';
+import { supabase } from '@/utils/supabase/client';
+import { v4 as uuidv4 } from 'uuid';
 
 export async function GET() {
   try {
@@ -46,5 +48,105 @@ export async function GET() {
   } catch (e) {
     console.error('Error fetching labs:', e);
     return Response.json({ error: 'Internal server error' }, { status: 500 });
+  }
+}
+
+//TODO: work on POST req Ui and how to do when google auth
+export async function POST(req: Request) {
+  try {
+    const body = await req.json();
+    const {
+      labName,
+      testType,
+      nextAvailable,
+      experienceYears,
+      imageUrl,
+      collectionTypes,
+      latitude,
+      longitude,
+      isAvailable,
+      isLoved,
+      rating,
+      timeSlots, // array of { time, session }
+    } = body;
+
+    const {
+      data: { user },
+      error,
+    } = await supabase.auth.getUser();
+
+    if (!user || error) {
+      return new Response(JSON.stringify({ error: 'Not authenticated' }), {
+        status: 401,
+      });
+    }
+
+    // Find lab belonging to this user
+    const lab = await db.lab.findUnique({ where: { userId: user.id } });
+    if (!lab) {
+      return new Response(JSON.stringify({ error: 'Lab not found' }), {
+        status: 404,
+      });
+    }
+
+    // Upsert LabDetails
+    await db.labDetails.upsert({
+      where: { labId: lab.id },
+      update: {
+        labName,
+        testType,
+        experienceYears: experienceYears ?? null,
+        imageUrl,
+        collectionTypes: Array.isArray(collectionTypes)
+          ? collectionTypes
+          : (collectionTypes?.split(',').map((s: string) => s.trim()) ?? []),
+        nextAvailable: nextAvailable ? new Date(nextAvailable) : null,
+        latitude: latitude ?? 0.0,
+        longitude: longitude ?? 0.0,
+        isLoved: isLoved ?? false,
+        isAvailable: isAvailable ?? true,
+        rating: rating ?? null,
+      },
+      create: {
+        id: uuidv4(),
+        labId: lab.id,
+        labName,
+        testType,
+        experienceYears: experienceYears ?? null,
+        imageUrl,
+        collectionTypes: Array.isArray(collectionTypes)
+          ? collectionTypes
+          : (collectionTypes?.split(',').map((s: string) => s.trim()) ?? []),
+        nextAvailable: nextAvailable ? new Date(nextAvailable) : null,
+        latitude: latitude ?? 0.0,
+        longitude: longitude ?? 0.0,
+        isLoved: isLoved ?? false,
+        isAvailable: isAvailable ?? true,
+        rating: rating ?? null,
+      },
+    });
+
+    // Reset and insert timeSlots
+    if (timeSlots && timeSlots.length > 0) {
+      await db.labTimeSlot.deleteMany({ where: { labId: lab.id } });
+      await db.labTimeSlot.createMany({
+        data: timeSlots.map((slot: { time: string; session: string }) => ({
+          id: uuidv4(),
+          labId: lab.id,
+          time: slot.time,
+          session: slot.session.toUpperCase(), // MORNING / AFTERNOON / EVENING
+        })),
+      });
+    }
+
+    return new Response(
+      JSON.stringify({ message: 'Lab details updated successfully' }),
+      { status: 201 }
+    );
+  } catch (error) {
+    console.error(error);
+    return new Response(JSON.stringify({ error: 'Internal server error' }), {
+      status: 500,
+    });
   }
 }
